Change dir structure
Created views dir and created more files for tab bar
