+++ packages/js/app/src/views/SettingsScreen.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Text, View } from 'react-native';
+
+
+const SettingsScreen = () => (
+    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+        <Text>Settings!</Text>
+    </View>
+);
+
+export default SettingsScreen;
